packages/backend-api/requirements.txt
fastapi[all]
uvicorn
pymongo
google-generativeai
azure-identity
azure-keyvault-secrets
python-dotenv
pytest
httpx